Created a new namespace: Parallel, this is under the Threading namespace.
The classes in this namespace offer you another way call delegates and reflectively call methods asynchronously.

Changed NotQuiteRight to old under the Items directory.

Added AssemblySafetyAttribute which can be used to indicate whether or not the assembly uses anything in an unsafe context.

There is some funny business with the settings files under properties.