Event, ActionEvent and FuncEvent have been added which can be used to call a list of delegates asynchronously to provide some kind of notification of something, to something.
Also added is a new type of abstract SubThread called BaseThreadWorker which is similar to the base isolated threads in that the classes must be
derived from to be used but differ in that multiple threads can be executed using the same ThreadMain method and can be passed parameters more or
less directly depending on the variant you choose.

ConcurrentValueContainer has a new property called UsesMemoryBarrier, when set to true a memory fence is issued by the contained Spinlock.

NullTolerantAttribute has been added which enables you to indicate that whatever it is the attribute has been applied to can handle null and
won't throw an exception if it is provided or assigned it.

TempReference has been renamed to DisposableReference, which is more appropriate considering the struct implements IDisposable.

Other minor changes have been made...