Added:
TieredConcurrentValueContainer - A ConcurrentValueContainer variant that has properties and methods that try to enter a Monitor lock before entering the spinlock. This feature is significant because the Monitor lock acts like a flow valve for the spinlock.

LockHog - Use this if you want to enter a Monitor lock and "hog" it.

AsyncLockHog - Use this if you want to hog a Monitor lock but don't want it to block the thread you're instigating it from as well.

ActiveWaitEvent - A wait event that spinwaits and checks a Monitor locked Boolean field to determine whether or not it has been signaled.

ConcurrentStringBuilder - Wraps a System.Text.StringBuilder instance and allows access through a Monitor lock.

ICloneWith - Indicates that the implementer clones itself but is affected by a parameter.

IToStringBuilder - Indicates that the implementer returns a StringBuilder with its contents in a text format of some kind.

Updated:

StringBuilderPool - Added a Remove method which removes a set number or instances from the StringBuilder queue.

BaseEvent - Added HasAddAction and HasRemoveAction properties.