Added:
CoreComponents:

DynamicAndNot - Contains a normal and dynamic reference of a DynamicObject inheritor so the dynamic and non-dynamic capabilities of the object can be conveniently used in tandem.

ExperimentalAttribute - Use this to decorate experimental classes, properties, methods etc.

VersionAttribute - Use this to decorate a member or an object with a version number.

CoreComponents.Text:

IAppend - An interface for appending to StringBuilders.

CoreComponents.Threading:

Clicker - Used to signal when an condition is true.

WaitableClicker - A Clicker variant that can be waited on.

ComplicatedClicker - A Clicker that can signal whether or not an exception has been caught.

EggTimer - Times out after a certain period of time. Once started you check on it or wait for it in the intervening time period.

TieredEggTimer - An EggTimer with a set of properties and methods which use Monitor locks for high intensity environments.

New namespace:

CoreComponents.W3Etc

This namespace contains objects for working with XML and related web technologies.

BaseElement, MultiTagElement, SingleTagElement, XComment and XDocument enable convenient creation and manipulation of XML and similar documents.

HTMLElements - Contains methods for creating single and multi tag HTML elements.

CSSDocument, Rule - Use these classes to quickly create CSS style rules and documents.

Changed:

IndentationLevelBuilder, SpaceIndentationLevelBuilder and TabIndentationLevelBulider have been changed to structs.

StringBuilderPool - Now has async methods

ConcurrentStringBulder - Has new AppendLine overload methods for Appending all of the types of objects that the Append overloads have prior to appending the newline characters.