Added:
CouldBeNull and ReadonlyCouldBeNull - Reference containers that can help enable a more defensive programming style.

IReferenceValueContainer and IReadonlyReferenceValueContainer - Interfaces for reference value containers.

ConcurrentCache and ConcurrentStringBuilderCache - Caching objects based on the System.Collections.Concurrent.ConcurrentQueue

IWaitable - For objects that wait

IWaitableValueContainer - For objects that wait and contain things

SpinWaitableValueContainer - Spin-waits when you interact the contained item and has a ManualResetEventSlim for when you want to wait for the value to be set.

QuickInputOutputMessage - Somewhat similar to the SpinWaitableValueContainer but also has an output variable aswell

MonitorJob - Like the Job but using Monitor locks

Changes:

Moved MonitorStringBuilder to Text

and a few other things